### 2024-03-22 #6526 @editorialbot
[PRE REVIEW]: CBX: Python and Julia Packages for Consensus-Based Interacting Particle Methods
**Submitting author:** @rafaelbailo (<a href="http://orcid.org/0000-0001-8018-3799">Rafael Bailo</a>)
**Repository:** https://github.com/PdIPS/CBX
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
