### 2024-03-22 #6526 @danielskatz
Submission message from the authors:
> This is a first submission to JOSS of two packages for consensus-based optimisation (unconstrained, global, derivative-free optimisation). This is a combined submission of two packages (one for Python, one for Julia) with common functionality. The Python package (https://pdips.github.io/CBXpy/) is hosted at https://github.com/PdIPS/CBXpy, and the Julia package (https://pdips.github.io/ConsensusBasedX.jl/) is hosted at https://github.com/PdIPS/ConsensusBasedX.jl . We are submitting the paper from a third repository, https://github.com/PdIPS/CBX, on the suggestion of Arfon Smith. A preprint of this paper has been made available on arXiv: https://arxiv.org/abs/2403.14470 . No portion of the paper will be submitted elsewhere.
