### 2024-03-22 c97ec55
Updated .bib to add DOIs
