+++ paper.bib
@@ -1,24 +1,3 @@
-@Article{pinnau2017consensus,
-  author    = {Pinnau, Ren{\'e} and Totzeck, Claudia and Tse, Oliver and Martin, Stephan},
-  title     = {A consensus-based model for global optimization and its mean-field limit},
-  number    = {01},
-  pages     = {183--204},
-  volume    = {27},
-  fjournal  = {Mathematical Models and Methods in Applied Sciences},
-  journal   = {Math. Models Methods Appl. Sci.},
-  publisher = {World Scientific},
-  year      = {2017},
-}
-
-@inproceedings{movckus1975bayesian,
-  title={On {B}ayesian methods for seeking the extremum},
-  author={Mo{\v{c}}kus, Jonas},
-  booktitle={Optimization Techniques IFIP Technical Conference: Novosibirsk, July 1--7, 1974},
-  pages={400--404},
-  year={1975},
-  organization={Springer}
-}
-
 @article{boltzmann1868studien,
   title={Studien ueber das {G}leichgewicht der lebenden {K}raft},
   author={Boltzmann, Ludwig},
@@ -27,37 +6,6 @@
   pages={49--96},
   year={1868},
   publisher={Barth}
-}
-
-@Article{mohan2012survey,
-  author    = {Mohan, B Chandra and Baskaran, R},
-  title     = {A survey: Ant Colony Optimization based recent research and implementation on several engineering domain},
-  number    = {4},
-  pages     = {4618--4627},
-  volume    = {39},
-  fjournal  = {Expert Systems with Applications},
-  journal   = {Expert Syst. Appl.},
-  publisher = {Elsevier},
-  year      = {2012},
-}
-
-@article{karaboga2014comprehensive,
-  title={A comprehensive survey: artificial bee colony (ABC) algorithm and applications},
-  author={Karaboga, Dervis and Gorkemli, Beyza and Ozturk, Celal and Karaboga, Nurhan},
-  journal={Artificial intelligence review},
-  volume={42},
-  pages={21--57},
-  year={2014},
-  publisher={Springer}
-}
-
-@inproceedings{yang2009firefly,
-  title={Firefly algorithms for multimodal optimization},
-  author={Yang, Xin-She},
-  booktitle={International symposium on stochastic algorithms},
-  pages={169--178},
-  year={2009},
-  organization={Springer}
 }
 
 @article{bayraktar2013wind,
@@ -147,14 +95,6 @@
   year      = {2022},
 }
 
-@book{ackley2012connectionist,
-  title={A connectionist machine for genetic hillclimbing},
-  author={Ackley, David},
-  volume={28},
-  year={2012},
-  publisher={Springer science \& business media}
-}
-
 @Article{harris2020array,
   author    = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
   title     = {Array programming with {NumPy}},
@@ -194,16 +134,6 @@
   journal   = {Comput. Sci. Eng.},
   publisher = {IEEE COMPUTER SOC},
   year      = {2007},
-}
-
-@article{miranda2018pyswarms,
[…]
