### 2024-03-22 #6526 @rafaelbailo
@mstimberg for Python, we suggest [Oliver Dunbar](https://reviewers.joss.theoj.org/reviewers/3343) and [Guillaume Dalle](https://reviewers.joss.theoj.org/reviewers/2816). For Julia, we suggest [Alessandro Pierro](https://reviewers.joss.theoj.org/reviewers/3342) and [Mathieu Besançon](https://reviewers.joss.theoj.org/reviewers/458).
