### 2024-04-08 #6526 @gdalle
Just to make sure that it is part of the public record as well (I have informed the editor): the last author of the paper @urbainvaes was a colleague of mine at École des Ponts. I left this institution very recently (2022) and will come back to it in 2025. However, we have never written anything together, and I don't believe this will affect my ability to judge the present paper.
