### 2024-04-08 #6526 @mstimberg
I confirm that we waive this potential conflict of interest in accordance with the [JOSS policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) (CC @danielskatz)
