### 2024-04-12 #6526 @mstimberg
Many thanks again to all reviewers (@Bobby-Huggins and @AlessandroPierro will focus on the Python package; @kellertuer and @gdalle will review on the Julia side) for agreeing to help out. I will now start the official review issue, all future discussions will take place over there :wave:
