### 2024-04-12 #6611 @editorialbot
[REVIEW]: CBX: Python and Julia Packages for Consensus-Based Interacting Particle Methods
**Submitting author:** @rafaelbailo (<a href="http://orcid.org/0000-0001-8018-3799">Rafael Bailo</a>)
**Repository:** https://github.com/PdIPS/CBX
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.6-v1.3.1

@AlessandroPierro & @gdalle & @kellertuer & @Bobby-Huggins, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
