### 2024-04-12 #6611 @mstimberg
👋🏼  @rafaelbailo @AlessandroPierro, @gdalle, @kellertuer, @Bobby-Huggins this is the review thread for the paper. All of our communications will happen here from now on.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/6611#issue-2240743891).

Please note that the repository linked above only contains the JOSS paper. The actual code is in
- https://github.com/PdIPS/CBXpy (`CBXpy`; python package)
- https://github.com/PdIPS/ConsensusBasedX.jl (`ConsensusBasedX.jl`; Julia package)

Since the JOSS machinery does not directly support this setup, please ignore the "Software report" with the lines of code above, and the "Failed to discover a valid open source license" statement (both packages are distributed under the OSI-approved MIT license).
