### 2024-04-16 #6611 @kellertuer
I have a short – more like organisational – question on one of the points to check, concerning

> Repository: Is the source code for this software available at the https://github.com/PdIPS/CBX?

The correct answer is: No. That is (just) the JOSS paper. But it is hard to specify what to link to since this paper concerns two software packages (https://github.com/PdIPS/CBXpy and https://github.com/PdIPS/ConsensusBasedX.jl) such that I feel the question is maybe in general not fitting completely here (unless one allows for two urls there).
