### 2024-04-16 PdIPS/CBX#3 @kellertuer
Review comments
Thanks for the nice paper and for submitting it to JOSS (cf https://github.com/openjournals/joss-reviews/issues/6611) . Here are a few comments both on the content of the rendered PDF of this repositories paper as well as on the Julia package. I will do them as check boxes but also number them in case you would like to comment / discuss. Once we agree they are finished, I can check them then.

## The paper

Line numbers in the following points refer to the version generated by the editorialbot on April 12, 2024 [here](https://github.com/PdIPS/CBX/issues/[https://raw.githubusercontent.com/openjournals/joss-papers/joss.06611/joss.06611/10.21105.joss.06611.pdf)
* [ ] 1. line 32ff. The authors have a nicve motivation and overview. For a unified presentation one could rewrite “where $\nabla f$ is not well-defined” also more like the part before “where f is nonsmooth ($\nabla f$ is not well-defined),or where the evaluation […]”
* [ ] 2. line 40: One could maybe mention that $t$ is time already in the sentence defining $x_t$ instead of in the next – but this might also still be fine as is.
* [ ] 3. line 99/100: a small remark for the statement of need. The authors state Optimization.jl also includes Nelder Mead. The other mentioned package (Manopt.jl) also includes Nelder-Mead and most recently (probably after submission of the authors) also CMA-ES, for which I am not sure how much this is related to the CBX family of solvers.
* [ ] 4. line 105: “easy-to-understand” is subjective, maybe one could just leave that out? I feel “high-level” is enough of an explanation that even beginner users should get an easy start.
* [ ] 5. It would maybe be nice – for example after the specifics – to do a short tutorial or a small example to illustrate how to use the package? If could also be within the “specifics”, then I would leave our that word, and add for both packages a small code snipped others could get started with. Since the Markdown is available, readers can even copy them, but a 5 line example or so would be nice to get started

## The Julia package

* [ ] 6. A contributing.md would maybe be nice explaining a bit how other users could contribute to the package.
* [ ] 7. I personally feel the aliases minimize/minimise/optimise/optimize are more confusing than helpful
* [ ] 8. The “Full code example” links in the documentation could also directly be put into the docs instead of just linked. For example as plain code blocks, but maybe even as Literate.jl or Quarto-rendered ones so the documentation also reports on the results.
* [ ] 9. Concerning tests – you have a quite high code coverage already – You could consider even reaching for a bit more, since it self the configuration calls can easily be tested; are for example the errors actually ever hit?. As for the parts in `utils/arrays.jl` maybe those are not necessary as functions when they are never called? In run-dynamics there is a few return nothing lines that seem to be never reached, are they necessary then?
* [x] 10. Since this is more a comment than a request, I already checked this one. Since you have quite a few citations in the docs, I felt it was great to use https://github.com/JuliaDocs/DocumenterCitations.jl in a documentation to have a way to cite papers.
