### 2024-04-16 #6611 @mstimberg
@kellertuer I agree that the situation is not ideal here, but unfortunately the JOSS system only allows for a single URL in a submission. Having a third repository with the JOSS paper in it was a suggestion by the editor-in-chief, so I think it is the best we can do for now. To make the checklists less ambiguous, I edited them to refer to the repository for either `CBXpy` or `ConsensusBasedX.jl` (in case you haven't seen it, I mentioned the URLs in [my earlier comment](https://github.com/openjournals/joss-reviews/issues/6611#issuecomment-2052389632)).

@rafaelbailo I noticed that the README at https://github.com/PdIPS/CBX only links to the documentation pages for  `CBXpy` and `ConsensusBasedX.jl`, could you please also include links to their GitHub repositories so that future visitors can easily find the source code?
