### 2024-04-16 #6611 @kellertuer
Sure, I of course found both easily, then I was maybe taking the question too rigorous (sorry, a German following rules here! ;) ).
I also think that the third repo with the paper is the best way when the paper is about two packages.
And with linking them in the readme I am also fine then.
