### 2024-04-19 e23d9f8
addressing reviewer comments in issue #3
