+++ paper.bib
@@ -19,6 +19,38 @@
   publisher={IEEE}
 }
 
+@misc{hansen2019pycma,
+  author       = {Nikolaus Hansen and Youhei Akimoto and Petr Baudis},
+  title        = {{CMA-ES/pycma} on {G}ithub},
+  howpublished = {Zenodo, DOI:10.5281/zenodo.2559634},
+  month        = feb,
+  year         = 2019,
+  doi          = {10.5281/zenodo.2559634},
+  url          = {https://doi.org/10.5281/zenodo.2559634},
+}
+
+@inproceedings{hansen1996adapting,
+  title={Adapting arbitrary normal mutation distributions in evolution strategies: The covariance matrix adaptation},
+  author={Hansen, Nikolaus and Ostermeier, Andreas},
+  booktitle={Proceedings of IEEE international conference on evolutionary computation},
+  pages={312--317},
+  year={1996},
+  organization={IEEE},
+  doi={10.1109/ICEC.1996.542381}
+}
+
+@article{colutto2009cma,
+  title={The CMA-ES on Riemannian manifolds to reconstruct shapes in 3-D voxel images},
+  author={Colutto, Sebastian and Fruhauf, Florian and Fuchs, Matthias and Scherzer, Otmar},
+  journal={IEEE Transactions on Evolutionary Computation},
+  volume={14},
+  number={2},
+  pages={227--245},
+  year={2009},
+  doi={10.1109/TEVC.2009.2029567},
+  publisher={IEEE}
+}
+
 @inproceedings{kennedy1995particle,
   title={Particle swarm optimization},
   author={Kennedy, James and Eberhart, Russell},
