### 2024-04-19 PdIPS/CBX#3 @TimRoith
Thank you very much for the comments! Since I am mainly concerned with the Python implementation, I will first only address the issues regarding the paper:

* Ad 1. We unified the presentation accordingly, and mimic the bracket structure from before.
* Ad 2. We introduce $t$ in the first sentence now.
* Ad 3. We now mention CMA-ES, especially we cite the Python and Julia implementations for it. The connection between CMA-ES and consensus based methods, is indeed interesting, however to the best of our knowledge, unexplored. In fact, it is part of ongoing work, therefore it is really important not forgetting to mention it here.
*Ad 4. We removed "easy-to-understand".
* Ad 5. For Python a small example code snippet is included, for Julia a similar one will follow.
