+++ paper.bib
@@ -168,6 +168,7 @@
   year      = {2007},
 }
 
+
 @software{scikitopt,
 author = {Guo, Fei},
 title = {{scikit-opt}},
