### 2024-04-19 af93a1b
added DOIs and tried eprint format
