+++ paper.bib
@@ -58,7 +58,8 @@
   volume={4},
   pages={1942--1948},
   year={1995},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/ICNN.1995.488968}
 }
 
 @article{henderson2003theory,
@@ -67,7 +68,8 @@
   journal={Handbook of metaheuristics},
   pages={287--319},
   year={2003},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/0-306-48056-5_10}
 }
 
 @article{rastrigin1963convergence,
@@ -96,13 +98,15 @@
   number={2},
   pages={212--229},
   year={1961},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/321062.321069}
 }
 
-@Article{fornasier2021consensus,
+@online{fornasier2021consensus,
   author  = {Fornasier, Massimo and Klock, Timo and Riedl, Konstantin},
   title   = {Consensus-based optimization methods converge globally},
-  journal = {Preprint arXiv:2103.15130},
+  eprint = {2103.15130},
+  eprinttype = {arXiv},
   year    = {2021},
 }
 
@@ -167,6 +171,7 @@
   publisher = {IEEE COMPUTER SOC},
   year      = {2007},
 }
+
 
 @software{scikitopt,
 author = {Guo, Fei},
