### 2024-04-19 9f0ea61
added more DOIs
