+++ paper.bib
@@ -16,7 +16,8 @@
   number={5},
   pages={2745--2757},
   year={2013},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TAP.2013.2238654}
 }
 
 @misc{hansen2019pycma,
@@ -116,7 +117,8 @@
   booktitle={Active Particles, Volume 3: Advances in Theory, Models, and Applications},
   pages={201--226},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-030-93302-9_6}
 }
 
 @Article{carrillo2022consensus,
@@ -129,6 +131,7 @@
   journal   = {Stud. Appl. Math.},
   publisher = {Wiley Online Library},
   year      = {2022},
+  doi       = {10.1111/sapm.12470}
 }
 
 @Article{harris2020array,
@@ -227,17 +230,19 @@ version = {1.0.1},
 year = {2023}
 }
 
-@article{duan2023pypop7,
+@online{duan2023pypop7,
   title={PyPop7: A Pure-Python Library for Population-Based Black-Box Optimization},
   author={Duan, Qiqi and Zhou, Guochen and Shao, Chang and Wang, Zhuowei and Feng, Mingyang and Yang, Yijun and Zhao, Qi and Shi, Yuhui},
-  journal={Preprint arXiv:2212.05652},
+  eprint={2212.05652},
+  eprinttype={arXiv},
   year={2022}
 }
 
-@Article{bungert2022polarized,
+@online{bungert2022polarized,
   author  = {Bungert, Leon and Wacker, Philipp and Roith, Tim},
   title   = {Polarized consensus-based dynamics for optimization and sampling},
-  journal = {Preprint arXiv:2211.05238},
+  eprint = {2211.05238},
+  eprinttype = {arXiv},
   year    = {2022},
 }
 
@@ -304,7 +309,8 @@ year = {2023}
   number={2},
   pages={30},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s00245-023-09983-3}
 }
 
 @Article{grassi2020particle,
@@ -322,17 +328,19 @@ year = {2023}
   year     = {2021},
 }
 
-@article{riedl2023gradient,
+@online{riedl2023gradient,
   title={Gradient is All You Need?},
   author={Riedl, Konstantin and Klock, Timo and Geldhauser, Carina and Fornasier, Massimo},
-  journal={Preprint arXiv:2306.09778},
+  eprint={2306.09778},
+  eprinttype = {arXiv},
   year={2023}
 }
 
-@article{carrillo2023fedcbo,
+@online{carrillo2023fedcbo,
   title={FedCBO: Reaching Group Consensus in Clustered Federated Learning through Consensus-based Optimization},
   author={Carrillo, Jos{\'e}~A and Garcia Trillos, Nicolas and Li, Sixu and Zhu, Yuhua},
-  journal={Preprint arXiv:2305.02894},
+  eprint={2305.02894},
+  eprinttype={arXiv},
   year={2023}
 }
 
@@ -344,7 +352,8 @@ year = {2023}
   number={2},
   pages={1--43},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  doi={doi.org/10.1007/s00245-023-10036-y}
 }
 
 @article{fornasier2020consensus_sphere_convergence,
@@ -372,6 +381,7 @@ MRREVIEWER = {Ctirad Matonoha},
   fjournal  = {SIAM Journal on Optimization},
   journal   = {SIAM J. Optim.},
   publisher = {SIAM},
+  doi       = {10.1137/22M1471304},
   year      = {2023},
 }
 
[…]
