+++ README.md
@@ -1,3 +1,3 @@
 # CBX: Python and Julia Packages for Consensus-Based Interacting Particle Methods
 
-This repository hosts the joint [JOSS](https://joss.theoj.org/) submission of the packages [CBXPy](https://pdips.github.io/CBXpy/) and [ConsensusBasedX.jl](https://pdips.github.io/ConsensusBasedX.jl/). It is authored by [Rafael Bailo](https://orcid.org/0000-0001-8018-3799), [Alethea Barbaro](https://orcid.org/0000-0001-9856-2818), [Susana N. Gomes](https://orcid.org/0000-0002-8731-367X), [Konstantin Riedl](https://orcid.org/0000-0002-2206-4334), [Tim Roith](https://orcid.org/0000-0001-8440-2928), [Claudia Totzeck](https://orcid.org/0000-0001-6283-7154), and [Urbain Vaes](https://orcid.org/0000-0002-7629-7184).
+This repository hosts the joint [JOSS](https://joss.theoj.org/) submission of the packages [CBXPy](https://github.com/PdIPS/CBXpy) and [ConsensusBasedX.jl](https://github.com/PdIPS/ConsensusBasedX.jl). It is authored by [Rafael Bailo](https://orcid.org/0000-0001-8018-3799), [Alethea Barbaro](https://orcid.org/0000-0001-9856-2818), [Susana N. Gomes](https://orcid.org/0000-0002-8731-367X), [Konstantin Riedl](https://orcid.org/0000-0002-2206-4334), [Tim Roith](https://orcid.org/0000-0001-8440-2928), [Claudia Totzeck](https://orcid.org/0000-0001-6283-7154), and [Urbain Vaes](https://orcid.org/0000-0002-7629-7184).
