### 2024-04-30 59b2b25
Julia minimal example
