### 2024-04-30 PdIPS/CBX#3 @rafaelbailo
Thank you for the comments @kellertuer! I believe points 5 and 6 are now addressed.

Regarding 7: with the exception of the note about the aliases, the documentation for CBO is written entirely in terms of `minimise`. The aliases can of course be ignored, but are there to increase compatibility with users from other communities: I have colleagues who always "minimise", and I have colleagues who always "optimise", and both groups can be quite vocal about their terminology. If it's alright with you, I would prefer to keep the aliases, just to avoid future headaches!
