### 2024-04-30 PdIPS/CBX#3 @rafaelbailo
Regarding 8: we considered using Literate.jl for the code examples to make the documentation easier to manage. However, we found that the examples are quite static (since the top-level interface is quite simple and very unlikely to change), and simultaneously we wanted to use the examples as part of the unit testing: https://github.com/PdIPS/ConsensusBasedX.jl/blob/main/test/examples.jl checks that every example can be run and generates no errors or warnings. Is there a good way to merge the testing and something like Literate.jl?
