### 2024-04-30 PdIPS/CBX#3 @rafaelbailo
Regarding 9: I've just pruned the superfluous methods and added a couple of additional unit tests. Our (free) codecov rate has maxed, but I will upload the new coverage report ASAP!
