### 2024-04-30 PdIPS/CBX#3 @kellertuer
> I have colleagues who always "minimise", and I have colleagues who always "optimise", and both groups can be quite vocal about their terminology.

I minimise by optimisation. ;)
I think my main complaint is really about _both_ American _and_ British English aliases. even without that I feel a package could have a consistent (that is one) terminology and other names could be mentioned in the docs.
