### 2024-04-30 PdIPS/CBX#3 @kellertuer
Concerning 8, one could also include the 3-4 lines just as code in the docs instead of putting them a click away.
I personally use Quarto, but Literate or Weave are not _too_ far way from that; still, avoiding a click to the example would already be the feature that would make me happy.

The remaining comments I will check the next days
