### 2024-05-04 #6611 @gdalle
I have complete my review, both the paper and package seem very high quality to me. I have made a few remarks in https://github.com/PdIPS/ConsensusBasedX.jl/issues/31, but my two main issues are:

- The tests don't seem to check for correctness of optimization, especially on the provided examples (but I may have missed that part)
- The documentation does not show the code examples
