### 2024-05-12 #6611 @AlessandroPierro
My review, focused on the Python side, is complete. The contribution is well polished and checks all JOSS requirements. I had a minor comment on contributing guidelines (ref. [PdIPS/CBXpy/issues/23](https://github.com/PdIPS/CBXpy/issues/23)) and the authors promptly addressed it.

@rafaelbailo, thank you for your submission!
