### 2024-05-14 PdIPS/CBX#3 @gdalle
I had not noticed this issue when I opened my own review issue here: https://github.com/PdIPS/ConsensusBasedX.jl/issues/31

Some of the concerns of @kellertuer are shared, so I'll jump in here.

>  I think my main complaint is really about both American and British English aliases. even without that I feel a package could have a consistent (that is one) terminology and other names could be mentioned in the docs.

Definitely agree with this one. Picking a side is not a bad thing.

> Concerning 8, one could also include the 3-4 lines just as code in the docs instead of putting them a click away.

Also a major agreement on my end.

>    > Is there a good way to merge the testing and something like Literate.jl?
>
> I am not aware of something doing that.

Not automatically, but you can use the same source files and include them both in docs and in tests. My package [HiddenMarkovModels.jl](https://github.com/gdalle/HiddenMarkovModels.jl) does that, although it might be slightly more complicated when you have several source files per docs page.

> https://github.com/PdIPS/ConsensusBasedX.jl/blob/main/test/examples.jl checks that every example can be run and generates no errors or warnings.

Are the examples ever checked for correctness? More generally, my [other major issue](https://github.com/PdIPS/ConsensusBasedX.jl/issues/31#issuecomment-2094108786) was that the tests never seem to verify that the result is the minimiser we expect
