### 2024-05-15 #6611 @Bobby-Huggins
My comments on the paper and Python package are here: https://github.com/PdIPS/CBXpy/issues/28

Overall excellent submission: I just echo @gdalle in mentioning that it looks like there could be more test coverage on the actual optimization results (but test coverage overall is great). I also mention one apparent bug in the plotting routine, and some small comments on the paper.
