+++ paper.bib
@@ -175,6 +175,47 @@
   year      = {2007},
 }
 
+@Misc{gpyopt2016,
+author = {The GPyOpt authors},
+title = {{GPyOpt}: A Bayesian Optimization framework in python},
+howpublished = {\url{http://github.com/SheffieldML/GPyOpt}},
+year = {2016}
+}
+
+@inproceedings{balandat2020botorch,
+  title={{BoTorch: A Framework for Efficient Monte-Carlo Bayesian Optimization}},
+  author={Balandat, Maximilian and Karrer, Brian and Jiang, Daniel R. and Daulton, Samuel and Letham, Benjamin and Wilson, Andrew Gordon and Bakshy, Eytan},
+  booktitle = {Advances in Neural Information Processing Systems 33},
+  year={2020},
+  url = {http://arxiv.org/abs/1910.06403}
+}
+
+@online{GPflowOpt2017,
+  author = {Knudde, Nicolas and {van der Herten}, Joachim and Dhaene, Tom and Couckuyt, Ivo},
+  title = "{{GP}flow{O}pt: {A} {B}ayesian {O}ptimization {L}ibrary using Tensor{F}low}",
+  eprint = {1711.03845},
+  eprinttype = {arXiv},
+  year    = {2017},
+  url     = {https://arxiv.org/abs/1711.03845}
+}
+
+@article{Jiménez2017, 
+doi = {10.21105/joss.00431}, 
+url = {https://doi.org/10.21105/joss.00431}, 
+year = {2017}, 
+publisher = {The Open Journal}, 
+volume = {2}, number = {19}, 
+pages = {431}, 
+author = {José Jiménez and Josep Ginebra}, 
+title = {pyGPGO: Bayesian Optimization for Python}, 
+journal = {Journal of Open Source Software} }
+
+@Misc{Bayesian14,
+    author = {Fernando Nogueira},
+    title = {{Bayesian Optimization}: Open source constrained global optimization tool for {Python}},
+    year = {2014--},
+    url = " https://github.com/bayesian-optimization/BayesianOptimization"
+}
 
 @software{scikitopt,
 author = {Guo, Fei},
