+++ paper.bib
@@ -308,6 +308,16 @@ year = {2023}
   doi = {10.5281/zenodo.7754236},
 }
 
+@article{Singh2024, 
+doi = {10.21105/joss.05694}, 
+url = {https://doi.org/10.21105/joss.05694}, 
+year = {2024}, publisher = {The Open Journal}, 
+volume = {9}, number = {94}, pages = {5694}, 
+author = {Gurjeet Sangra Singh and Luigi Acerbi}, 
+title = {PyBADS: Fast and robust black-box optimization in Python}, 
+journal = {Journal of Open Source Software} 
+}
+
 @Article{carrillo2018analytical,
   author   = {Carrillo, Jos{\'e}~A and Choi, Young-Pil and Totzeck, Claudia and Tse, Oliver},
   title    = {An analytical framework for consensus-based global optimization method},
