### 2024-05-31 #6611 @mstimberg
:wave: @rafaelbailo and @TimRoith, could you give us an estimate when you'll be able to address the remaining comments by the reviewers? As fars I can tell, PdIPS/ConsensusBasedX.jl#31 and PdIPS/CBX#3 are still waiting for changes/replies/clarifications. Thanks!
