### 2024-05-31 #6611 @gdalle
Yes indeed, thanks for answering my remarks about the paper in https://github.com/PdIPS/ConsensusBasedX.jl/issues/31, but my main concerns remain about the documentation and the tests.
