### 2024-06-06 PdIPS/CBX#3 @gdalle
It is much nicer for users, thank you!
Do the examples ever run during docs build or tests? Is there anything preventing them from getting outdated without you noticing?
