### 2024-06-06 b04bfcd
Fix citation inconsistencies
