+++ paper.bib
@@ -122,7 +122,7 @@
 }
 
 @Article{carrillo2022consensus,
-  author    = {Carrillo, Jos{\'e}~A and Hoffmann, Franca and Stuart, Andrew M and Vaes, Urbain},
+  author    = {Carrillo, J A and Hoffmann, Franca and Stuart, Andrew M and Vaes, Urbain},
   title     = {Consensus-based sampling},
   number    = {3},
   pages     = {1069--1140},
@@ -319,7 +319,7 @@ journal = {Journal of Open Source Software}
 }
 
 @Article{carrillo2018analytical,
-  author   = {Carrillo, Jos{\'e}~A and Choi, Young-Pil and Totzeck, Claudia and Tse, Oliver},
+  author   = {Carrillo, J A and Choi, Young-Pil and Totzeck, Claudia and Tse, Oliver},
   title    = {An analytical framework for consensus-based global optimization method},
   doi      = {10.1142/S0218202518500276},
   issn     = {0218-2025},
@@ -389,7 +389,7 @@ journal = {Journal of Open Source Software}
 
 @online{carrillo2023fedcbo,
   title={FedCBO: Reaching Group Consensus in Clustered Federated Learning through Consensus-based Optimization},
-  author={Carrillo, Jos{\'e}~A and Garcia Trillos, Nicolas and Li, Sixu and Zhu, Yuhua},
+  author={Carrillo, J A and Garcia Trillos, Nicolas and Li, Sixu and Zhu, Yuhua},
   eprint={2305.02894},
   eprinttype={arXiv},
   year={2023}
@@ -641,7 +641,7 @@ pages={1–24}}
 }
 
 @Article{carrillo2021consensus,
-  author       = {Carrillo, Jos{\'e} Antonio and Jin, Shi and Li, Lei and Zhu, Yuhua},
+  author       = {Carrillo, J A and Jin, Shi and Li, Lei and Zhu, Yuhua},
   date         = {2021},
   journaltitle = {ESAIM: Control, Optimisation and Calculus of Variations},
   title        = {A consensus-based global optimization method for high dimensional machine learning problems},
