### 2024-06-06 PdIPS/CBX#3 @rafaelbailo
Yes, the unit tests still run every single file in the examples folder, checking that the code executes without errors.
