### 2024-06-06 PdIPS/CBX#3 @gdalle
But is the result ever checked against what is expected? We can maybe discuss it over in my review since it was my main worry, don't want to hijack this one: https://github.com/PdIPS/ConsensusBasedX.jl/issues/31#issuecomment-2094108786
