### 2024-06-07 #6611 @rafaelbailo
Hi @mstimberg, we've just implemented the changes required in [PdIPS/ConsensusBasedX.jl#31](https://github.com/PdIPS/ConsensusBasedX.jl/issues/31) and [PdIPS/CBX#3](https://github.com/PdIPS/CBX/issues/3). Thank you for your patience (and to @kellertuer and @gdalle)!
