### 2024-06-07 PdIPS/CBX#3 @rafaelbailo
Sorry, I understood that the problem was the British/American alias:
> I minimise by optimisation. ;) I think my main complaint is really about _both_ American _and_ British English aliases. even without that I feel a package could have a consistent (that is one) terminology and other names could be mentioned in the docs.

Shall I remove `optimise` also?
