### 2024-06-07 PdIPS/CBX#3 @kellertuer
No, the problem was actually both.

I think a package should have exactly one language they use. Sure the documentation can state “People form this area might refer to A as B”. But if you provide 2 functions, why not also the German, French, Norwegian, Swedish,... translations?
So with “language” I meant both the different language aliases _as well as_ the different terminology aliases.

So please decide to use only one of those 4.
