### 2024-06-07 PdIPS/CBX#3 @kellertuer
Wrong choice. I should of course have even `optimise` to keep ;) No I am joking, that would just be _my_ terminology, but I am in no place to prescribe you a terminology.

Good that you narrowed it down to one.
