### 2024-06-07 c9184a4
Fix GPyOpy citation
