+++ paper.bib
@@ -176,7 +176,7 @@
 }
 
 @Misc{gpyopt2016,
-author = {The GPyOpt authors},
+author = {{The GPyOpt authors}},
 title = {{GPyOpt}: A Bayesian Optimization framework in python},
 howpublished = {\url{http://github.com/SheffieldML/GPyOpt}},
 year = {2016}
