### 2024-06-07 #6611 @rafaelbailo
@mstimberg I have a question. I've just changed affiliations, and I'm unsure of which affiliation to list on the paper. Is the JOSS policy to list the affiliation where most of the work was done, the most recent affiliation, or both?
