### 2024-06-07 #6611 @mstimberg
@rafaelbailo I am not aware that JOSS has a specific policy in that regard, meaning that it is up to you (or e.g. your employer's policy). But I will check back with the other editors to be sure.
