### 2024-06-10 #6611 @mstimberg
Hi again @rafaelbailo. I checked back with the other editors, and there is indeed no specific policy from JOSS's side about this. In some fields, it would be most common to list the affiliation where you did your work (and maybe add a footnote about your changed affiliation). But others would consider it most common to use the current affiliation (and potentially add a footnote in the other sense). Note that in either case, if you keep your ORCID page up to date, both affiliations will be linked via it. So to sum up: please use your best judgement, we are not imposing either convention from our side :blush:
