### 2024-06-17 #6611 @gdalle
Apart from the missing/invalid DOIs and some last remarks in https://github.com/PdIPS/ConsensusBasedX.jl/issues/31, this is good to go on my end
