### 2024-06-17 #6611 @kellertuer
Besides the remarks from @gdalle, the only thing I noticed is that now the Julia and Python packages are inconsistent - one uses optimise (upon my request to not use 4 names for the same function) and the python one uses minimise for the same – has the Python package more names than that?
