### 2024-06-17 #6611 @TimRoith
I guess right now, Python (only) uses optimize and Julia minimise.
