### 2024-06-17 #6611 @kellertuer
If it really only defines that in Python that is fine with me.
