### 2024-06-17 #6611 @gdalle
@TimRoith now Figure 1 is missing altogether. I guess that is one way of making it not blurry
