### 2024-06-17 #6611 @TimRoith
Either I am missing something or the template does not support pdf embedding
