+++ paper.bib
@@ -109,6 +109,7 @@
   eprint = {2103.15130},
   eprinttype = {arXiv},
   year    = {2021},
+  doi = {10.48550/arXiv.2103.15130}
 }
 
 @incollection{totzeck2021trends,
@@ -196,7 +197,8 @@ year = {2016}
   eprint = {1711.03845},
   eprinttype = {arXiv},
   year    = {2017},
-  url     = {https://arxiv.org/abs/1711.03845}
+  url     = {https://arxiv.org/abs/1711.03845},
+  doi     = {https://doi.org/10.48550/arXiv.1711.03845}
 }
 
 @article{Jiménez2017, 
@@ -276,15 +278,18 @@ year = {2023}
   author={Duan, Qiqi and Zhou, Guochen and Shao, Chang and Wang, Zhuowei and Feng, Mingyang and Yang, Yijun and Zhao, Qi and Shi, Yuhui},
   eprint={2212.05652},
   eprinttype={arXiv},
-  year={2022}
+  year={2022},
+  doi={10.48550/arXiv.2212.05652}
 }
 
-@online{bungert2022polarized,
-  author  = {Bungert, Leon and Wacker, Philipp and Roith, Tim},
-  title   = {Polarized consensus-based dynamics for optimization and sampling},
-  eprint = {2211.05238},
-  eprinttype = {arXiv},
-  year    = {2022},
+@ARTICLE{bungert2022polarized,
+  title     = "Polarized consensus-based dynamics for optimization and sampling",
+  author    = "Bungert, Leon and Roith, Tim and Wacker, Philipp",
+  journal   = "Math. Program.",
+  publisher = "Springer Science and Business Media LLC",
+  month     =  may,
+  year      =  2024,
+  doi       = {10.1007/s10107-024-02095-y}
 }
 
 @Article{Bergmann2022,
@@ -384,7 +389,8 @@ journal = {Journal of Open Source Software}
   author={Riedl, Konstantin and Klock, Timo and Geldhauser, Carina and Fornasier, Massimo},
   eprint={2306.09778},
   eprinttype = {arXiv},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2306.09778}
 }
 
 @online{carrillo2023fedcbo,
@@ -392,7 +398,8 @@ journal = {Journal of Open Source Software}
   author={Carrillo, J A and Garcia Trillos, Nicolas and Li, Sixu and Zhu, Yuhua},
   eprint={2305.02894},
   eprinttype={arXiv},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2305.02894}
 }
 
 @article{borghi2022adaptive,
@@ -404,7 +411,7 @@ journal = {Journal of Open Source Software}
   pages={1--43},
   year={2023},
   publisher={Springer},
-  doi={doi.org/10.1007/s00245-023-10036-y}
+  doi={10.1007/s00245-023-10036-y}
 }
 
 @article{fornasier2020consensus_sphere_convergence,
@@ -485,7 +492,8 @@ pages={1–24}}
   author={Althaus, Konstantin and Papaioannou, Iason and Ullmann, Elisabeth},
   eprint={2304.09077},
   eprinttype={arXiv},
-  year={2023}
+  year={2023},
+  doi={https://doi.org/10.48550/arXiv.2304.09077}
 }
 
 @Misc{DR2023,
