+++ paper.bib
@@ -198,12 +198,12 @@ year = {2016}
   eprinttype = {arXiv},
   year    = {2017},
   url     = {https://arxiv.org/abs/1711.03845},
-  doi     = {https://doi.org/10.48550/arXiv.1711.03845}
+  doi     = {10.48550/arXiv.1711.03845}
 }
 
 @article{Jiménez2017, 
 doi = {10.21105/joss.00431}, 
-url = {https://doi.org/10.21105/joss.00431}, 
+url = {10.21105/joss.00431}, 
 year = {2017}, 
 publisher = {The Open Journal}, 
 volume = {2}, number = {19}, 
