+++ paper.bib
@@ -27,7 +27,7 @@
   month        = feb,
   year         = 2019,
   doi          = {10.5281/zenodo.2559634},
-  url          = {https://doi.org/10.5281/zenodo.2559634},
+  url          = {10.5281/zenodo.2559634},
 }
 
 @inproceedings{hansen1996adapting,
@@ -315,7 +315,7 @@ year = {2023}
 
 @article{Singh2024, 
 doi = {10.21105/joss.05694}, 
-url = {https://doi.org/10.21105/joss.05694}, 
+url = {10.21105/joss.05694}, 
 year = {2024}, publisher = {The Open Journal}, 
 volume = {9}, number = {94}, pages = {5694}, 
 author = {Gurjeet Sangra Singh and Luigi Acerbi}, 
@@ -493,7 +493,7 @@ pages={1–24}}
   eprint={2304.09077},
   eprinttype={arXiv},
   year={2023},
-  doi={https://doi.org/10.48550/arXiv.2304.09077}
+  doi={10.48550/arXiv.2304.09077}
 }
 
 @Misc{DR2023,
