### 2024-06-18 #6611 @TimRoith
Sending regards up north, hope the weather is better than in rainy Hamburg ;)
