### 2024-06-18 #6611 @kellertuer
It‘s actually ok with not too much rain today.

What are the next steps now, are we the only two (Myself and Guillaume) that were missing?
