### 2024-06-19 #6611 @mstimberg
Thanks @TimRoith. I am in contact with other editors to figure out how to best handle the fact that this submission has two packages. The main obstacle is that the version number that we define here has to correspond to a tag in the submission repository (https://github.com/PdIPS/CBX), and should correspond to the code that is the outcome of the review process (it will get archived). One solution that I see is that you add CBXPy and ConsensusBasedX.jl as git submodules to the CBX repository. These can then both point to the latest respective release, and you can tag the CBX repository with a "joint tag" like "v0.1.6-v1.3.0" or similar. Would that work for you?
