### 2024-06-19 #6611 @rafaelbailo
I've created a new release of ConsensusBasedX.jl, [v1.3.1](https://github.com/PdIPS/ConsensusBasedX.jl/releases/tag/v1.3.1). Its Zenodo archive is [![DOI](https://zenodo.org/badge/614934085.svg)](https://zenodo.org/doi/10.5281/zenodo.12169479).

Together with the release of CBXpy, both have been added as submodules of the paper repository, and a [v0.1.6-v1.3.1](https://github.com/PdIPS/CBX/releases/tag/v0.1.6-v1.3.1) release has been made, with Zenodo archive [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.12169935.svg)](https://doi.org/10.5281/zenodo.12169935).

I've also double-checked the author lists and licenses, so I think we're good to go!
