### 2024-06-20 #6611 @mstimberg
Thanks @rafaelbailo. Unfortunately, I just realized that the `git archive` created by GitHub (and used by Zenodo) does not contain the content of submodules (see zenodo/zenodo#1036). This means that the zip file in your Zenodo archive contains two empty folders for the code... Could you please prepare a new archive that includes the full code (if you used the Zenodo-GitHub bridge you will instead have to do it manually)? Either upload three separate zip files (CBX, CBXPy and ConsensusBased.jl), or create a single archive with something like this:
```
git archive -o release.zip v0.1.6-v1.3.1
git submodule --quiet foreach 'cd $toplevel; zip -ru release.zip $sm_path'
```
Also, please change the title of the archive so that it matches the title of the paper. In the meantime, I will go through the paper and let you know whether any changes are necessary there ­– please note that changes to the paper do not require the creation of a new archive, so you won't have to do it again.
