### 2024-06-20 860f2ef
minor typographic fixes for bibliography
