+++ paper.bib
@@ -41,7 +41,7 @@
 }
 
 @article{colutto2009cma,
-  title={The CMA-ES on Riemannian manifolds to reconstruct shapes in 3-D voxel images},
+  title={The CMA-ES on {R}iemannian manifolds to reconstruct shapes in 3-D voxel images},
   author={Colutto, Sebastian and Fruhauf, Florian and Fuchs, Matthias and Scherzer, Otmar},
   journal={IEEE Transactions on Evolutionary Computation},
   volume={14},
@@ -92,7 +92,7 @@
 }
 
 @article{hooke1961direct,
-  title={"Direct Search" Solution of Numerical and Statistical Problems},
+  title={``Direct Search'' Solution of Numerical and Statistical Problems},
   author={Hooke, Robert and Jeeves, Terry A},
   journal={Journal of the ACM (JACM)},
   volume={8},
@@ -170,21 +170,20 @@
   abstract  = {Matplotlib is a 2D graphics package used for {P}ython for
   application development, interactive scripting, and publication-quality
   image generation across user interfaces and operating systems.},
-  fjournal  = {Computing in Science & Engineering},
-  journal   = {Comput. Sci. Eng.},
+  journal   = {Computing in Science \& Engineering},
   publisher = {IEEE COMPUTER SOC},
   year      = {2007},
 }
 
 @Misc{gpyopt2016,
 author = {{The GPyOpt authors}},
-title = {{GPyOpt}: A Bayesian Optimization framework in python},
+title = {GPyOpt: A {B}ayesian Optimization framework in {P}ython},
 howpublished = {\url{http://github.com/SheffieldML/GPyOpt}},
 year = {2016}
 }
 
 @inproceedings{balandat2020botorch,
-  title={{BoTorch: A Framework for Efficient Monte-Carlo Bayesian Optimization}},
+  title={BoTorch: A Framework for Efficient Monte-Carlo {B}ayesian Optimization},
   author={Balandat, Maximilian and Karrer, Brian and Jiang, Daniel R. and Daulton, Samuel and Letham, Benjamin and Wilson, Andrew Gordon and Bakshy, Eytan},
   booktitle = {Advances in Neural Information Processing Systems 33},
   year={2020},
@@ -193,7 +192,7 @@ year = {2016}
 
 @online{GPflowOpt2017,
   author = {Knudde, Nicolas and {van der Herten}, Joachim and Dhaene, Tom and Couckuyt, Ivo},
-  title = "{{GP}flow{O}pt: {A} {B}ayesian {O}ptimization {L}ibrary using Tensor{F}low}",
+  title = "{GPflowOpt: A {B}ayesian Optimization Library using TensorFlow}",
   eprint = {1711.03845},
   eprinttype = {arXiv},
   year    = {2017},
@@ -209,13 +208,13 @@ publisher = {The Open Journal},
 volume = {2}, number = {19}, 
 pages = {431}, 
 author = {José Jiménez and Josep Ginebra}, 
-title = {pyGPGO: Bayesian Optimization for Python}, 
+title = {pyGPGO: {B}ayesian Optimization for {P}ython},
 journal = {Journal of Open Source Software} }
 
 @Misc{Bayesian14,
     author = {Fernando Nogueira},
     title = {{Bayesian Optimization}: Open source constrained global optimization tool for {Python}},
-    year = {2014--},
+    year = {2014–},
     url = " https://github.com/bayesian-optimization/BayesianOptimization"
 }
 
@@ -259,7 +258,7 @@ year = {2021},
 
 @software{Igor_CBOinPython,
 author = {Tukh, Igor and Riedl, Konstantin},
-title = {{cbo-in-python}},
+title = {{CBO-in-python}},
[…]
