### 2024-06-20 a8bde2c
Expand journal abbreviations
