+++ paper.bib
@@ -128,8 +128,7 @@
   number    = {3},
   pages     = {1069--1140},
   volume    = {148},
-  fjournal  = {Studies in Applied Mathematics},
-  journal   = {Stud. Appl. Math.},
+  journal   = {Studies in Applied Mathematics},
   publisher = {Wiley Online Library},
   year      = {2022},
   doi       = {10.1111/sapm.12470}
@@ -155,8 +154,7 @@
   pages    = {261--272},
   volume   = {17},
   adsurl   = {https://rdcu.be/b08Wh},
-  fjournal = {Nature Methods},
-  journal  = {Nat. Methods},
+  journal  = {Nature Methods},
   year     = {2020},
 }
 
@@ -208,7 +206,7 @@ publisher = {The Open Journal},
 volume = {2}, number = {19}, 
 pages = {431}, 
 author = {José Jiménez and Josep Ginebra}, 
-title = {pyGPGO: {B}ayesian Optimization for {P}ython},
+title = {pyGPGO: {B}ayesian Optimization for {P}ython}, 
 journal = {Journal of Open Source Software} }
 
 @Misc{Bayesian14,
@@ -242,8 +240,7 @@ year = {2021},
   title    = {{DEAP}: Evolutionary Algorithms Made Easy},
   pages    = {2171--2175},
   volume   = {13},
-  fjournal = {Journal of Machine Learning Research},
-  journal  = {J. Mach. Learn. Res.},
+  journal  = {Journal of Machine Learning Research},
   month    = {7},
   year     = {2012},
 }
@@ -318,7 +315,7 @@ url = {10.21105/joss.05694},
 year = {2024}, publisher = {The Open Journal}, 
 volume = {9}, number = {94}, pages = {5694}, 
 author = {Gurjeet Sangra Singh and Luigi Acerbi}, 
-title = {PyBADS: Fast and robust black-box optimization in {P}ython},
+title = {PyBADS: Fast and robust black-box optimization in {P}ython}, 
 journal = {Journal of Open Source Software} 
 }
 
@@ -330,8 +327,7 @@ journal = {Journal of Open Source Software}
   number   = {6},
   pages    = {1037--1066},
   volume   = {28},
-  fjournal = {Mathematical Models and Methods in Applied Sciences},
-  journal  = {Math. Models Methods Appl. Sci.},
+  journal  = {Mathematical Models and Methods in Applied Sciences},
   mrclass  = {90C26 (35Q84 37N40 60H10 90C59)},
   mrnumber = {3804923},
   year     = {2018},
@@ -350,8 +346,7 @@ journal = {Journal of Open Source Software}
   title     = {Leveraging memory effects and gradient information in consensus-based optimisation: On global convergence in mean-field law},
   doi       = {10.1017/S0956792523000293},
   pages     = {1–32},
-  fjournal  = {European Journal of Applied Mathematics},
-  journal   = {European J. Appl. Math.},
+  journal   = {European Journal of Applied Mathematics},
   publisher = {Cambridge University Press},
   year      = {2023},
 }
@@ -376,8 +371,7 @@ journal = {Journal of Open Source Software}
   number   = {8},
   pages    = {1625--1657},
   volume   = {31},
-  fjournal = {Mathematical Models and Methods in Applied Sciences},
-  journal  = {Math. Models Methods Appl. Sci.},
+  journal  = {Mathematical Models and Methods in Applied Sciences},
   mrclass  = {35Q83 (65K10 65M75 82C40)},
   mrnumber = {4307005},
   year     = {2021},
@@ -418,8 +412,7 @@ journal = {Journal of Open Source Software}
               S\"{u}nnen, Philippe},
      TITLE = {Consensus-based optimization on the sphere: convergence to
[…]
