### 2024-06-20 eca1762
Fix Boltzmann (1868) reference
