+++ paper.bib
@@ -1,11 +1,10 @@
 @article{boltzmann1868studien,
-  title={Studien ueber das {G}leichgewicht der lebenden {K}raft},
+  title={Studien über das {G}leichgewicht der lebendigen {K}raft zwischen bewegten materiellen {P}unkten},
   author={Boltzmann, Ludwig},
-  journal={Wissenschafiliche Abhandlungen},
-  volume={1},
-  pages={49--96},
-  year={1868},
-  publisher={Barth}
+  journal={Wiener Berichte},
+  volume={58},
+  pages={517--560},
+  year={1868}
 }
 
 @article{bayraktar2013wind,
