### 2024-06-21 #6611 @mstimberg
> Hi @mstimberg, I've followed your instructions and created a new archive: https://zenodo.org/records/12207224 .

Thanks @rafaelbailo. Looks very good to me, just two minor corrections (you should be able to edit the metadata manually, no need to create a new archive):
- could you please set the license to MIT license – it currently states CC-By
- could you set the "version" displayed on Zenodo to `v0.1.6-v1.3.1` (it is currently listed as `v3`).

After that, we should be all good from my side and I will hand over to an EiC for the final editing/publishing steps.
