### 2024-06-21 #6611 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
