### 2024-06-21 bfdf0cb
very minor change in paper.md
