### 2024-06-21 #6611 @rafaelbailo
I've just been told by @TimRoith that his DESY affiliation cannot be edited, the institution requires that it's exactly as was before the changes of @mstimberg . Can we revert it please?
