### 2024-06-21 592e62c
minor changes in paper.bib
