+++ paper.bib
@@ -1,5 +1,5 @@
 @article{boltzmann1868studien,
-  title={Studien über das {G}leichgewicht der lebendigen {K}raft zwischen bewegten materiellen {P}unkten},
+  title={Studien über das Gleichgewicht der lebendigen Kraft zwischen bewegten materiellen Punkten},
   author={Boltzmann, Ludwig},
   journal={Wiener Berichte},
   volume={58},
@@ -21,7 +21,7 @@
 
 @misc{hansen2019pycma,
   author       = {Nikolaus Hansen and Youhei Akimoto and Petr Baudis},
-  title        = {{CMA-ES/pycma} on {G}ithub},
+  title        = {{CMA-ES/pycma} on {G}it{H}ub},
   howpublished = {Zenodo, DOI:10.5281/zenodo.2559634},
   month        = feb,
   year         = 2019,
@@ -180,7 +180,7 @@ year = {2016}
 }
 
 @inproceedings{balandat2020botorch,
-  title={BoTorch: A Framework for Efficient Monte-Carlo {B}ayesian Optimization},
+  title={BoTorch: A Framework for Efficient {M}onte-{C}arlo {B}ayesian Optimization},
   author={Balandat, Maximilian and Karrer, Brian and Jiang, Daniel R. and Daulton, Samuel and Letham, Benjamin and Wilson, Andrew Gordon and Bakshy, Eytan},
   booktitle = {Advances in Neural Information Processing Systems 33},
   year={2020},
@@ -269,7 +269,7 @@ year = {2023}
 }
 
 @online{duan2023pypop7,
-  title={PyPop7: A Pure-Python Library for Population-Based Black-Box Optimization},
+  title={PyPop7: A Pure-{P}ython Library for Population-Based Black-Box Optimization},
   author={Duan, Qiqi and Zhou, Guochen and Shao, Chang and Wang, Zhuowei and Feng, Mingyang and Yang, Yijun and Zhao, Qi and Shi, Yuhui},
   eprint={2212.05652},
   eprinttype={arXiv},
@@ -568,7 +568,6 @@ pages={1–24}}
   pages        = {169--178},
   publisher    = {Springer Berlin Heidelberg},
   issn         = {1611-3349},
-  organization = {Springer},
   year         = {2009},
 }
 
