### 2024-06-21 #6611 @danielskatz
@rafaelbailo - I've proofread the paper, and suggest the changes in https://github.com/PdIPS/CBX/pull/5 - please merge this. Also, please go ahead and change the affiliation as you mention above. Then let me know, and we'll generate another proof to make sure everything is ok, then publish.
