### 2024-06-21 #6611 @mstimberg
Yes, please go ahead and commit the fix. Changes like this are still possible. They only become a bit of a hassle after @danielskatz issues the final accept command.
