+++ paper.bib
@@ -1,5 +1,5 @@
 @article{boltzmann1868studien,
-  title={Studien über das Gleichgewicht der lebendigen Kraft zwischen bewegten materiellen Punkten},
+  title={Studien über das {G}leichgewicht der lebendigen {K}raft zwischen bewegten materiellen {P}unkten},
   author={Boltzmann, Ludwig},
   journal={Wiener Berichte},
   volume={58},
