### 2024-06-21 #6611 @danielskatz
@TimRoith - please check this proof and confirm it is ok with you. I will also check it.
