### 2024-06-21 #6611 @danielskatz
Congratulations to @rafaelbailo (Rafael Bailo) and co-authors on your publication!

And thanks to @AlessandroPierro, @gdalle, @kellertuer, and @Bobby-Huggins for reviewing,
and to @mstimberg for editing!
JOSS depends on volunteers and couldn't succeed without you
